+++ lib/maru/builder/methods.ex
@@ -6,6 +6,7 @@ defmodule Maru.Builder.Methods do
   alias Maru.Struct.Resource
   alias Maru.Struct.Parameter
   alias Maru.Struct.Endpoint
+  alias Maru.Struct.Route
   alias Maru.Struct.Plug, as: MaruPlug
   alias Maru.Builder.Path, as: MaruPath
 
@@ -15,8 +16,8 @@ defmodule Maru.Builder.Methods do
     @doc "Handle #{method} method."
     defmacro unquote(method)(path \\ "", [do: block]) do
       %{ method: unquote(method) |> to_string |> String.upcase,
-         path: path |> MaruPath.split,
-         block: block |> Macro.escape,
+         path:   path |> MaruPath.split,
+         block:  block |> Macro.escape,
        } |> endpoint
     end
   end
@@ -24,8 +25,8 @@ defmodule Maru.Builder.Methods do
   @doc "Handle all method."
   defmacro match(path \\ "", [do: block]) do
     %{ method: Macro.var(:_, nil) |> Macro.escape,
-       path: path |> MaruPath.split,
-       block: block |> Macro.escape,
+       path:   path |> MaruPath.split,
+       block:  block |> Macro.escape,
      } |> endpoint
   end
 
@@ -37,17 +38,22 @@ defmodule Maru.Builder.Methods do
           [] else [{:version}]
         end
       @endpoints %Endpoint{
+        func_id: @func_id,
+        block:   unquote(ep.block)
+      }
+      @routes %Route{
         desc:       @desc,
         method:     unquote(ep.method),
         version:    resource.version,
         path:       version ++ resource.path ++ unquote(ep.path),
         parameters: resource.parameters ++ Parameter.pop,
         helpers:    resource.helpers,
-        block:      unquote(ep.block),
         plugs:      MaruPlug.merge(resource.plugs, MaruPlug.pop),
-        __file__:   __ENV__.file,
+        module:     __MODULE__,
+        func_id:    @func_id,
       }
       @desc nil
+      @func_id @func_id + 1
     end
   end