+++ lib/maru/builder/plugs/server.ex
@@ -7,17 +7,17 @@ defmodule Maru.Builder.Plugs.Server do
   def compile(%Macro.Env{module: module}=env) do
     config      = (Application.get_env(:maru, module) || [])[:versioning] || []
     extend_opts = Module.get_attribute(module, :extend)
-    endpoints   = Module.get_attribute(module, :endpoints) ++
+    routes      = Module.get_attribute(module, :routes) ++
                   Module.get_attribute(module, :mounted)
-    extended    = Maru.Builder.Extend.take_extended(endpoints, extend_opts)
-    endpoints   = endpoints ++ extended
+    extended    = Maru.Builder.Extend.take_extended(routes, extend_opts)
+    routes      = routes ++ extended
 
     adapter = Maru.Builder.Versioning.get_adapter(config[:using])
 
     pipeline = [
       Module.get_attribute(module, :plugs_before) |> Enum.reverse,
       adapter.plug(config),
-      [ {:endpoint, [], true},
+      [ {:route, [], true},
         {Maru.Plugs.NotFound, [], true},
       ],
     ] |> Enum.concat |> Enum.reverse
@@ -30,24 +30,24 @@ defmodule Maru.Builder.Plugs.Server do
       |> Enum.map(&Maru.Builder.Exceptions.make_rescue_block/1)
       |> List.flatten
 
-    endpoints_block =
-      Enum.map(endpoints, fn ep ->
-        Maru.Builder.Endpoint.dispatch(ep, env, adapter)
+    routes_block =
+      Enum.map(routes, fn route ->
+        Maru.Builder.Route.dispatch(route, env, adapter)
       end)
 
     method_not_allow_block =
-      Enum.group_by(endpoints, fn ep -> {ep.version, ep.path} end)
-      |> Enum.map(fn {{version, path}, endpoints} ->
-        unless Enum.any?(endpoints, fn i -> i.method == {:_, [], nil} end) do
-          Maru.Builder.Endpoint.dispatch_405(version, path, adapter)
+      Enum.group_by(routes, fn route -> {route.version, route.path} end)
+      |> Enum.map(fn {{version, path}, routes} ->
+        unless Enum.any?(routes, fn i -> i.method == {:_, [], nil} end) do
+          Maru.Builder.Route.dispatch_405(version, path, adapter)
         end
       end)
 
     quote do
-      unquote(endpoints_block)
+      unquote(routes_block)
       unquote(method_not_allow_block)
 
-      defp endpoint(conn, _), do: conn
+      defp route(conn, _), do: conn
 
       def init(_), do: []