+++ lib/maru/builder/plugs/test.ex
@@ -5,7 +5,7 @@ defmodule Maru.Builder.Plugs.Test do
 
   @doc false
   def compile(%Macro.Env{module: module}=env) do
-    endpoints = Module.get_attribute(module, :endpoints)
+    routes = Module.get_attribute(module, :routes)
     adapter   = Maru.Builder.Versioning.Test
 
     exceptions =
@@ -14,30 +14,30 @@ defmodule Maru.Builder.Plugs.Test do
       |> Enum.map(&Maru.Builder.Exceptions.make_rescue_block/1)
       |> List.flatten
 
-    endpoints_block =
-      Enum.map(endpoints, fn ep ->
-        Maru.Builder.Endpoint.dispatch(ep, env, adapter)
+    routes_block =
+      Enum.map(routes, fn route ->
+        Maru.Builder.Route.dispatch(route, env, adapter)
       end)
 
     method_not_allow_block =
-      Enum.group_by(endpoints, fn ep -> {ep.version, ep.path} end)
-      |> Enum.map(fn {{version, path}, endpoints} ->
-        unless Enum.any?(endpoints, fn i -> i.method == {:_, [], nil} end) do
-          Maru.Builder.Endpoint.dispatch_405(version, path, adapter)
+      Enum.group_by(routes, fn ep -> {ep.version, ep.path} end)
+      |> Enum.map(fn {{version, path}, routes} ->
+        unless Enum.any?(routes, fn i -> i.method == {:_, [], nil} end) do
+          Maru.Builder.Route.dispatch_405(version, path, adapter)
         end
       end)
 
     quote do
-      unquote(endpoints_block)
+      unquote(routes_block)
       unquote(method_not_allow_block)
 
-      defp endpoint_test(conn, _), do: conn
+      defp route_test(conn, _), do: conn
 
       def init(_), do: []
 
       def call_test(conn, _) do
         conn
-        |> endpoint_test([])
+        |> route_test([])
         |> case do
           %Plug.Conn{halted: true} = conn -> conn
           %Plug.Conn{}             = conn -> Maru.Plugs.NotFound.call(conn, [])