+++ lib/mix/tasks/routes.ex
@@ -48,18 +48,18 @@ defmodule Mix.Tasks.Maru.Routes do
 
     adapter = Maru.Builder.Versioning.get_adapter(config[:using])
 
-    module.__endpoints__
-    |> Enum.sort_by(fn ep -> ep.version end)
-    |> Enum.map(&generate_endpoint(&1, adapter, prefix))
+    module.__routes__
+    |> Enum.sort_by(fn route -> route.version end)
+    |> Enum.map(&generate_route(&1, adapter, prefix))
     |> Enum.map(&IO.puts/1)
   end
 
-  defp generate_endpoint(ep, adapter, prefix) do
-    path = adapter.path_for_params(ep.path, ep.version)
-    format_version(ep.version) <>
-    format_method(ep.method) <>
+  defp generate_route(route, adapter, prefix) do
+    path = adapter.path_for_params(route.path, route.version)
+    format_version(route.version) <>
+    format_method(route.method) <>
     format_path(prefix ++ path) <>
-    "  " <> (ep.desc || "")
+    "  " <> (route.desc || "")
   end
 
   defp format_version(nil), do: "nil" |> format_version