+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   * totally rewrite params parsing logic
   * add overridable plug
   * support top-level plug
+  * split Route and Endpoint
 
 ## v0.9.5 (2016-3-15)
 * Enhancements